move notes-search-list to /views/notes-page and update routing
