+++ app/src/app/app-routing-stix.module.ts
@@ -20,7 +20,7 @@ import { NgModule } from '@angular/core';
 import { environment } from "../environments/environment"
 import { AuthorizationGuard } from "./services/helpers/authorization.guard";
 import { Role } from "./classes/authn/role";
-import { NotesSearchListComponent } from "./views/stix/notes-search/notes-search-list/notes-search-list.component";
+import { NotesSearchListComponent } from "./views/notes-page/notes-search-list/notes-search-list.component";
 
 var viewRoles = [Role.VISITOR, Role.EDITOR, Role.ADMIN];
 var editRoles = [Role.EDITOR, Role.ADMIN];
@@ -498,23 +498,6 @@ const stixRoutes: Routes = [{
         ]
     }
     ]
-  },{
-    path: 'notes',
-    canActivateChild: [AuthorizationGuard],
-    data: {
-      breadcrumb: 'notes',
-      more: true
-    },
-    children: [{
-        path: '',
-        data: {
-          breadcrumb: 'list',
-          title: "Notes Search",
-          roles: viewRoles
-        },
-        component: NotesSearchListComponent
-      }
-    ]
   }
 ]
 
@@ -622,6 +605,28 @@ stixRoutes.push(
         ]
     }
 )
+
+
+stixRoutes.push(
+  {
+    path: 'notes',
+    canActivateChild: [AuthorizationGuard],
+    data: {
+      breadcrumb: 'notes',
+      more: true
+    },
+    children: [{
+        path: '',
+        data: {
+          breadcrumb: 'list',
+          title: "Notes Search",
+          roles: editRoles
+        },
+        component: NotesSearchListComponent
+      }
+    ]
+  }
+);
 
 @NgModule({
     imports: [
