+++ app/src/app/app.module.ts
@@ -187,7 +187,7 @@ import { MarkingDefinitionListComponent } from "./views/stix/marking-definition/
 import { CampaignViewComponent } from './views/stix/campaign/campaign-view/campaign-view.component';
 import { CampaignListComponent } from './views/stix/campaign/campaign-list/campaign-list.component';
 
-import { NotesSearchListComponent } from './views/stix/notes-search/notes-search-list/notes-search-list.component';
+import { NotesSearchListComponent } from './views/notes-page/notes-search-list/notes-search-list.component';
 
 @NgModule({
   declarations: [
