+++ app/src/app/views/notes-page/notes-search-list/notes-search-list.component.html
@@ -1,0 +1,18 @@
+<div class="list-page">
+  <div class="list-page-header">
+      <h1 class="with-help-icon">
+          notes
+          <popover-content #objectDescription placement="bottom">
+            <span>Notes allow users to add additional information about an object in the dataset without extending it directly. They are associated with objects within the knowledge base and cannot exist without an attached object. Notes can be searched by title or content and viewed in the context of the object they are associated with.</span>
+          </popover-content>
+          <mat-icon class="help-icon" [popover]="objectDescription" aria-label="help">help_outline</mat-icon>
+      </h1>
+  </div>
+  <app-stix-list [config]="{
+    'type': 'note',
+    'showLinks': true,
+    'showFilters': false,
+    'clickBehavior': 'linkToObjectRef'
+  }
+  "></app-stix-list>
+</div>
