+++ app/src/app/views/notes-page/notes-search-list/notes-search-list.component.ts
@@ -1,5 +1,4 @@
 import { Component, OnInit } from '@angular/core';
-import { SidebarService } from 'src/app/services/sidebar/sidebar.service';
 
 @Component({
   selector: 'app-notes-search-list',
@@ -7,11 +6,9 @@ import { SidebarService } from 'src/app/services/sidebar/sidebar.service';
   styleUrls: ['./notes-search-list.component.scss']
 })
 export class NotesSearchListComponent implements OnInit {
-  constructor(private sidebarService: SidebarService) { 
+  constructor() { 
   }
 
-  ngOnInit(): void {
-    this.sidebarService.opened = false;
-  }
+  ngOnInit(): void {}
 
 }
