--- app/src/app/views/stix/notes-search/notes-search-list/notes-search-list.component.html
@@ -1,20 +1,0 @@
-<div class="list-page">
-  <div class="list-page-header">
-      <h1 class="with-help-icon">
-          notes 
-          <mat-icon class="help-icon" aria-label="help"
-                      popover="Search for notes by name or content"
-                      popoverPlacement="bottom"
-                      >
-              help_outline
-          </mat-icon>
-      </h1>
-  </div>
-  <app-stix-list [config]="{
-    'type': 'note',
-    'showLinks': true,
-    'showFilters': false,
-    'clickBehavior': 'linkToObjectRef'
-  }
-  "></app-stix-list>
-</div>
